moving Change to common and fixing imports
