+++ src/providers/firestore.ts
@@ -26,12 +26,12 @@ import * as _ from 'lodash';
 import { posix } from 'path';
 import { apps } from '../apps';
 import {
-  Change,
   CloudFunction,
   Event,
   EventContext,
   makeCloudFunction,
 } from '../cloud-functions';
+import { Change } from '../common/core';
 import { dateToTimestampProto } from '../encoder';
 import { DeploymentOptions } from '../function-configuration';
 import * as logger from '../logger';
